requirements.txt
tensorflow-gpu==1.2.1
tensorflow-cpu==1.2.1
numpy
